+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 4.4.1-dev
+
 ## 4.4.0
 
 * Mention how the `allocator` argument relates to imports in the `DartEmitter`